run go fmt, minor fixes
